backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
pydantic==2.5.2
python-multipart==0.0.6
websockets==12.0
langchain==0.0.339
tenacity==8.2.3
redis==5.0.1
minio==7.2.0 